New doc generator, merge styletron-client and styletron-client-ie9
