+++ documentation.yml
@@ -1,0 +1,25 @@
+toc:
+  - name: Introduction
+    file: README.md
+  - name: styletron-core
+    file: packages/styletron-core/README.md
+  - StyletronCore
+  - name: styletron-client
+    file: packages/styletron-client/README.md
+  - StyletronClient
+  - StyletronClientIE9
+  - name: styletron-server
+    file: packages/styletron-server/README.md
+  - StyletronServer
+  - name: styletron-react
+    file: packages/styletron-react/README.md
+  - styledReact
+  - StyletronProviderReact
+  - name: styletron-preact
+    file: packages/styletron-preact/README.md
+  - styledPreact
+  - StyletronProviderPreact
+  - name: styletron-inferno
+    file: packages/styletron-inferno/README.md
+  - styledInferno
+  - StyletronProviderInferno
