--- jsdoc.json
@@ -1,34 +1,0 @@
-{
-    "tags": {
-        "allowUnknownTags": true
-    },
-    "source": {
-        "include": [
-            "./README.md",
-            "packages/styletron-core/src",
-            "packages/styletron-client/src",
-            "packages/styletron-server/src",
-            "packages/styletron-react/src",
-            "packages/styletron-inferno/src",
-            "packages/styletron-preact/src"
-        ],
-        "includePattern": ".js$"
-    },
-    "plugins": [
-        "plugins/markdown"
-    ],
-    "opts": {
-        "template": "node_modules/@rtsao/docdash",
-        "encoding": "utf8",
-        "destination": "htmldocs/",
-        "recurse": true,
-        "verbose": true
-    },
-    "templates": {
-        "cleverLinks": false,
-        "monospaceLinks": false
-    },
-    "docdash": {
-        "sort": false
-    }
-}
