--- packages/styletron-client-ie9/README.md
@@ -1,3 +1,0 @@
-# styletron-client-ie9
-
-Styletron client with support for IE9 stylesheet rule limits.
