+++ packages/styletron-preact/src/styled.js
@@ -7,7 +7,8 @@ export default styled;
 
 /**
  * Helper function to create styled element components
- * @packagename styletron-preact
+ * @name styledPreact
+ * @exportName styled
  * @param  {String|function} base     Tag name or styled element component
  * @param  {function|object} styleFn  Style object or function that returns a style object
  * @return {function}                 Styled element component
