+++ packages/styletron-react/src/provider.js
@@ -3,10 +3,17 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 
+import type StyletronCore from 'styletron-core';
+
+type providerPropsT = {|
+  styletron: StyletronCore,
+  children: React.Element<any>,
+|};
+
 /**
- * @class StyletronProvider
- * @packagename styletron-react
- * @description Provides a Styletron instance to descendant styled components via context
+ * Provides a Styletron instance to descendant styled components via context
+ * @name StyletronProviderReact
+ * @exportName StyletronProvider
  * @example
  * const Styletron = require('styletron');
  *
@@ -22,14 +29,6 @@ import PropTypes from 'prop-types';
  * @property {ReactElement} children - children
  * @extends ReactClass
  */
-
-import type StyletronCore from 'styletron-core';
-
-type providerPropsT = {|
-  styletron: StyletronCore,
-  children: React.Element<any>,
-|};
-
 class StyletronProvider extends React.Component<providerPropsT> {
   styletron: StyletronCore;
